refactor(seances): extraire SeanceCard de TeacherSeances (#28)
Tranche 2 de la décomposition de TeacherSeances.vue.

- SeanceCard.vue (567 l., <script setup>) : sous-composant de présentation de la
  carte de séance (états visio aucune/programmée/active/terminée). Props
  seance/isEnseignant/actionLoading ; émet activate/start/deactivate/join/end.
  Helpers formatDate/formatTime locaux + CSS scopée copiée à l'identique.
- La vue utilise <SeanceCard v-for> ; markup carte + CSS + formatters morts
  retirés (1307 -> 769 l.). Les handlers visio (services, visioParticipations,
  buildJitsiUrl) restent au parent ; la carte émet.

Cumul TeacherSeances : 1354 -> 769 l. (-43 %).

Dette tracée : SeanceCard 567 l. (surtout CSS) dépasse §1.1. Parité visuelle :
CSS à l'identique + build vert, QA manuelle recommandée (pas de test UI).

Tests 261/261, contract 28/28, build OK.
